Fix UserDTO and UserDTOBuilder
