+++ src/test/java/com/thonecardoso/bookstoremanager/users/builder/UserDTOBuilder.java
@@ -1,5 +1,7 @@
 package com.thonecardoso.bookstoremanager.users.builder;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.thonecardoso.bookstoremanager.users.dto.UserDTO;
 import com.thonecardoso.bookstoremanager.users.enums.Gender;
 import lombok.Builder;
 
@@ -18,18 +20,32 @@ public class UserDTOBuilder {
     private int age = 31;
 
     @Builder.Default
-    private Gender gender;
+    private Gender gender = Gender.MALE;
 
     @Builder.Default
-    private String email;
+    private String email = "thone@teste.com";
 
     @Builder.Default
-    private String username;
+    private String username = "thonecardoso";
 
     @Builder.Default
-    private String password;
+    private String password = "123456";
 
     @Builder.Default
-    private LocalDate birthdate;
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd/MM/yyyy")
+    private LocalDate birthdate = LocalDate.of(1989, 8, 21);
+
+    public UserDTO buildUserDTO(){
+        return new UserDTO(
+                id,
+                name,
+                age,
+                gender,
+                email,
+                username,
+                password,
+                birthdate
+        );
+    }
 
 }
